Refactoring: dotcomponents -> dotcomrendering
